requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
requests==2.32.3
paho-mqtt==1.6.1
markdown==3.7
pydantic==2.9.2
ffmpeg-python==0.2.0
python-frontmatter==1.0.0
